+++ test_project/blog/migrations/0003_post_user.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+from django.conf import settings
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
+        ('blog', '0002_tag'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='post',
+            name='user',
+            field=models.ForeignKey(default=None, to=settings.AUTH_USER_MODEL),
+            preserve_default=False,
+        ),
+    ]
+