+++ test_project/test_migrations/unoptimized_initial_migration/blog_0001_project.py
@@ -9,14 +9,14 @@ class Migration(migrations.Migration):
     replaces = [('blog', '0001_initial'), ('blog', '0002_tag')]
 
     dependencies = [
-        ('cookbook', '0002_cookware'),
+        ('cookbook', '0001_project'),
     ]
 
     operations = [
         migrations.CreateModel(
             name='Post',
             fields=[
-                ('id', models.AutoField(primary_key=True, auto_created=True, serialize=False, verbose_name='ID')),
+                ('id', models.AutoField(serialize=False, primary_key=True, auto_created=True, verbose_name='ID')),
                 ('title', models.CharField(max_length=32)),
                 ('recipe', models.ForeignKey(to='cookbook.Recipe')),
             ],
@@ -27,7 +27,7 @@ class Migration(migrations.Migration):
         migrations.CreateModel(
             name='Tag',
             fields=[
-                ('id', models.AutoField(primary_key=True, auto_created=True, serialize=False, verbose_name='ID')),
+                ('id', models.AutoField(serialize=False, primary_key=True, auto_created=True, verbose_name='ID')),
                 ('name', models.CharField(max_length=32)),
                 ('posts', models.ManyToManyField(to='blog.Post')),
             ],