+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog (django-migrate-project)
 ==================================
 
+0.2.0 (Oct 10, 2015)
+--------------------
+
+- Renamed main commands to 'collectmigrations' and 'applymigrations'
+- Added experimental functionality for project-level migrations
+- Tightened up collect migrations functionality and added more complex tests
+- Fixed broken builds
+
 0.1.0 (May 15, 2015)
 --------------------